+++ PIR/include/ExecutionEngine.h
@@ -16,6 +16,7 @@
 
 #include "./Module.h"
 #include "../src/ExecutionEngineExecutor.h"
+#include "./JITEngine.h"
 
 
 namespace pcit::pir{
@@ -25,15 +26,19 @@ namespace pcit::pir{
 	class ExecutionEngine{
 		public:
 			using FuncRunError = ExecutionEngineExecutor::FuncRunError;
+			using InitConfig = JITEngine::InitConfig;
 
 		public:
-			#if defined(EVO_PLATFORM_WINDOWS)
-				ExecutionEngine(Module& _module) : module(_module) {}
-				~ExecutionEngine() = default;
-			#else
-				ExecutionEngine(Module& _module);
-				~ExecutionEngine();
-			#endif
+			ExecutionEngine(Module& _module);
+			~ExecutionEngine();
+
+
+			// if returns error, not initialized
+			// error is list of messages from LLVM
+			EVO_NODISCARD auto init(const InitConfig& config) -> evo::Expected<void, evo::SmallVector<std::string>>;
+			EVO_NODISCARD auto isInitialized() const -> bool { return this->jit_engine.isInitialized(); }
+
+
 
 			auto runFunction(Function::ID func_id, std::span<core::GenericValue> arguments)
 				-> evo::Expected<core::GenericValue, FuncRunError>;
@@ -43,12 +48,33 @@ namespace pcit::pir{
 			using Executor = ExecutionEngineExecutor;
 			EVO_NODISCARD auto get_current_executor() -> Executor&;
 
+
+			struct LoweredResult{
+				bool needs_to_be_lowered;
+				std::atomic<bool>& was_finished_being_lowered;
+			};
+			EVO_NODISCARD auto check_global_lowered(GlobalVar::ID global_id) -> LoweredResult {
+				const auto lock = std::scoped_lock(this->lowered_lock);
+
+				const auto find = this->lowered_globals.find(global_id);
+				if(find != this->lowered_globals.end()){ return LoweredResult(false, find->second); }
+
+				std::atomic<bool>& new_lowered_flag = this->finished_lowered_flags.emplace_back();
+				this->lowered_globals.emplace(global_id, new_lowered_flag);
+				return LoweredResult(true, new_lowered_flag);
+			}
+
 		private:
 			Module& module;
 
 			evo::StepVector<Executor> executors_alloc{};
 			std::unordered_map<std::thread::id, Executor&> executors{};
 			mutable evo::SpinLock executors_lock{};
+
+			JITEngine jit_engine{};
+			evo::StepVector<std::atomic<bool>> finished_lowered_flags{};
+			std::unordered_map<GlobalVar::ID, std::atomic<bool>&> lowered_globals{};
+			mutable evo::SpinLock lowered_lock{};
 
 			friend class ExecutionEngineExecutor;
 			friend void pcit::pir::_internal_signal_handler(int);